output: Extract file review model selections
File-review model construction had to derive actionable selections before
it could segment rows into review changes. That derivation handles batch
ownership groups, reset atoms, and display ID mapping.

Keeping that policy in the model builder made the builder responsible for
preparing selections and for shaping ReviewChange records.

This commit moves selection derivation into
output.file_review_model_selections and has the model builder call that
helper before row segmentation.

Subsequent commits will add architecture coverage for the model-selection
boundary.
